Restructure api machinery
