--- thoth-manager/src/api/models/mod.rs
@@ -1,8 +1,0 @@
-pub mod contribution;
-pub mod contributor;
-pub mod imprint;
-pub mod issue;
-pub mod publication;
-pub mod publisher;
-pub mod series;
-pub mod work;
