+++ thoth-manager/src/component/contributions_form.rs
@@ -9,17 +9,17 @@ use yewtil::fetch::FetchState;
 use yewtil::future::LinkFuture;
 use yewtil::NeqAssign;
 
-use crate::api::contribution_types_query::FetchActionContributionTypes;
-use crate::api::contribution_types_query::FetchContributionTypes;
-use crate::api::contributors_query::ContributorsRequest;
-use crate::api::contributors_query::ContributorsRequestBody;
-use crate::api::contributors_query::FetchActionContributors;
-use crate::api::contributors_query::FetchContributors;
-use crate::api::contributors_query::Variables;
-use crate::api::contributors_query::CONTRIBUTORS_QUERY;
-use crate::api::models::contribution::Contribution;
-use crate::api::models::contribution::ContributionTypeValues;
-use crate::api::models::contributor::Contributor;
+use crate::models::contribution::contribution_types_query::FetchActionContributionTypes;
+use crate::models::contribution::contribution_types_query::FetchContributionTypes;
+use crate::models::contributor::contributors_query::ContributorsRequest;
+use crate::models::contributor::contributors_query::ContributorsRequestBody;
+use crate::models::contributor::contributors_query::FetchActionContributors;
+use crate::models::contributor::contributors_query::FetchContributors;
+use crate::models::contributor::contributors_query::Variables;
+use crate::models::contributor::contributors_query::CONTRIBUTORS_QUERY;
+use crate::models::contribution::Contribution;
+use crate::models::contribution::ContributionTypeValues;
+use crate::models::contributor::Contributor;
 use crate::component::utils::FormBooleanSelect;
 use crate::component::utils::FormContributionTypeSelect;
 use crate::component::utils::FormTextInput;
