+++ thoth-manager/src/component/contributor.rs
@@ -11,13 +11,13 @@ use crate::agent::notification_bus::NotificationBus;
 use crate::agent::notification_bus::NotificationDispatcher;
 use crate::agent::notification_bus::NotificationStatus;
 use crate::agent::notification_bus::Request;
-use crate::api::models::contributor::Contributor;
-use crate::api::contributor_query::FetchActionContributor;
-use crate::api::contributor_query::FetchContributor;
-use crate::api::contributor_query::Variables;
-use crate::api::contributor_query::CONTRIBUTOR_QUERY;
-use crate::api::contributor_query::ContributorRequest;
-use crate::api::contributor_query::ContributorRequestBody;
+use crate::models::contributor::Contributor;
+use crate::models::contributor::contributor_query::FetchActionContributor;
+use crate::models::contributor::contributor_query::FetchContributor;
+use crate::models::contributor::contributor_query::Variables;
+use crate::models::contributor::contributor_query::CONTRIBUTOR_QUERY;
+use crate::models::contributor::contributor_query::ContributorRequest;
+use crate::models::contributor::contributor_query::ContributorRequestBody;
 use crate::component::utils::FormTextInput;
 use crate::component::utils::FormUrlInput;
 use crate::component::utils::Loader;
