+++ thoth-manager/src/component/issues_form.rs
@@ -7,14 +7,14 @@ use yewtil::fetch::FetchState;
 use yewtil::future::LinkFuture;
 use yewtil::NeqAssign;
 
-use crate::api::serieses_query::SeriesesRequest;
-use crate::api::serieses_query::SeriesesRequestBody;
-use crate::api::serieses_query::FetchActionSerieses;
-use crate::api::serieses_query::FetchSerieses;
-use crate::api::serieses_query::Variables;
-use crate::api::serieses_query::SERIESES_QUERY;
-use crate::api::models::issue::Issue;
-use crate::api::models::series::Series;
+use crate::models::issue::Issue;
+use crate::models::series::Series;
+use crate::models::series::serieses_query::SeriesesRequest;
+use crate::models::series::serieses_query::SeriesesRequestBody;
+use crate::models::series::serieses_query::FetchActionSerieses;
+use crate::models::series::serieses_query::FetchSerieses;
+use crate::models::series::serieses_query::Variables;
+use crate::models::series::serieses_query::SERIESES_QUERY;
 use crate::component::utils::FormNumberInput;
 use crate::string::EMPTY_ISSUES;
 use crate::string::REMOVE_BUTTON;
