+++ thoth-manager/src/component/work.rs
@@ -14,19 +14,19 @@ use crate::agent::notification_bus::NotificationBus;
 use crate::agent::notification_bus::NotificationDispatcher;
 use crate::agent::notification_bus::NotificationStatus;
 use crate::agent::notification_bus::Request;
-use crate::api::models::contribution::Contribution;
-use crate::api::models::imprint::Imprint;
-use crate::api::models::issue::Issue;
-use crate::api::models::publication::Publication;
-use crate::api::models::work::Work;
-use crate::api::models::work::WorkStatusValues;
-use crate::api::models::work::WorkTypeValues;
-use crate::api::work_query::FetchActionWork;
-use crate::api::work_query::FetchWork;
-use crate::api::work_query::Variables;
-use crate::api::work_query::WorkRequest;
-use crate::api::work_query::WorkRequestBody;
-use crate::api::work_query::WORK_QUERY;
+use crate::models::contribution::Contribution;
+use crate::models::imprint::Imprint;
+use crate::models::issue::Issue;
+use crate::models::publication::Publication;
+use crate::models::work::Work;
+use crate::models::work::WorkStatusValues;
+use crate::models::work::WorkTypeValues;
+use crate::models::work::work_query::FetchActionWork;
+use crate::models::work::work_query::FetchWork;
+use crate::models::work::work_query::Variables;
+use crate::models::work::work_query::WorkRequest;
+use crate::models::work::work_query::WorkRequestBody;
+use crate::models::work::work_query::WORK_QUERY;
 use crate::component::contributions_form::ContributionsFormComponent;
 use crate::component::issues_form::IssuesFormComponent;
 use crate::component::publications_form::PublicationsFormComponent;
