+++ thoth-manager/src/component/works.rs
@@ -9,14 +9,14 @@ use yewtil::fetch::FetchAction;
 use yewtil::fetch::FetchState;
 use yewtil::future::LinkFuture;
 
-use crate::api::models::contribution::Contribution;
-use crate::api::models::work::Work;
-use crate::api::works_query::FetchActionWorks;
-use crate::api::works_query::FetchWorks;
-use crate::api::works_query::Variables;
-use crate::api::works_query::WORKS_QUERY;
-use crate::api::works_query::WorksRequest;
-use crate::api::works_query::WorksRequestBody;
+use crate::models::contribution::Contribution;
+use crate::models::work::Work;
+use crate::models::work::works_query::FetchActionWorks;
+use crate::models::work::works_query::FetchWorks;
+use crate::models::work::works_query::Variables;
+use crate::models::work::works_query::WORKS_QUERY;
+use crate::models::work::works_query::WorksRequest;
+use crate::models::work::works_query::WorksRequestBody;
 use crate::component::utils::Loader;
 use crate::component::utils::Reloader;
 use crate::route::AdminRoute;
