+++ thoth-manager/src/models/contribution/contribution_types_query.rs
@@ -1,7 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::api::models::contribution::ContributionTypeDefinition;
+use super::ContributionTypeDefinition;
 
 const CONTRIBUTION_TYPES_QUERY: &str = "
     {
