+++ thoth-manager/src/models/series/serieses_query.rs
@@ -1,7 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::api::models::series::Series;
+use super::Series;
 
 pub const SERIESES_QUERY: &str = "
     query SeriesesQuery($filter: String) {
