+++ thoth-manager/src/models/stats/mod.rs
@@ -1,0 +1,1 @@
+pub mod stats_query;
