+++ thoth-manager/src/models/stats/stats_query.rs
